pyproject.toml
[project]
name = "macro-equity-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "fredapi>=0.5.2",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "yfinance>=1.2.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.5",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/macro_pipeline"]